# M10/tms/tms-api/contract/openapi.yaml
openapi: 3.0.3
info:
  title: TMS API
  description: Transportation Management System API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local environment
...
paths:
...
  /drivers:
    get:
      summary: List drivers
      description: Returns a list of all drivers (up to 100). No pagination.
      operationId: getDrivers
      tags:
        - Drivers
      responses:
        '200':
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriverListItem'
              example:
                - id: 1
                  first_name: Matt
                  last_name: Jerde
                  email: celiawalker@swift.net
                  phone: '4111234962'
                  contract_type: CONTRACTOR
                  status: ON_ROUTE
                - id: 2
                  first_name: Mekhi
                  last_name: Konopelski
                  email: karsonschaefer@franecki.info
                  phone: '6392038976'
                  contract_type: CONTRACTOR
                  status: RESTING
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a driver
      description: Creates a new driver record. All fields are optional.
      operationId: createDriver
      tags:
        - Drivers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverCreateInput'
            example:
              first_name: Jan
              last_name: Kowalski
              email: jan.kowalski@example.com
              phone: '1234567890'
              contract_type: FULL_TIME
              status: ACTIVE
      responses:
        '201':
          description: Driver created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverListItem'
              example:
                id: 21
                first_name: Jan
                last_name: Kowalski
                email: jan.kowalski@example.com
                phone: '1234567890'
                contract_type: FULL_TIME
                status: ACTIVE
        '400':
          description: Request payload has an invalid structure or contains missing/invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drivers/{id}:
    get:
      summary: Get driver by ID
      description: Returns a single driver with their associated driver licenses.
      operationId: getDriverById
      tags:
        - Drivers
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      responses:
        '200':
          description: Driver found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverDetail'
              example:
                id: 1
                first_name: Matt
                last_name: Jerde
                email: celiawalker@swift.net
                phone: '4111234962'
                contract_type: CONTRACTOR
                status: ON_ROUTE
                licenses:
                  - id: 1
                    document_number: CXX232105
                    issue_date: '2020-08-30'
                    expiry_date: '2027-09-21'
                    status: active
                    code: C
                    name: Prawo jazdy kat. C
                    description: Pojazdy ciężarowe powyżej 3,5t
                  - id: 2
                    document_number: C+E728699
                    issue_date: '2023-01-07'
                    expiry_date: '2028-01-07'
                    status: active
                    code: C+E
                    name: Prawo jazdy kat. C+E
                    description: Ciągnik siodłowy z naczepą
        '400':
          description: The provided ID is not a valid positive integer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No driver exists with the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Driver not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transportation-orders:
    get:
      summary: List transportation orders
      description: Returns a list of transportation orders, optionally filtered by customer ID. Results are ordered by order_date descending, limited to 100 items.
      operationId: getTransportationOrders
      tags:
        - TransportationOrders
      parameters:
        - name: customer_id
          in: query
          description: Filter orders by customer ID. When omitted, returns all orders.
          required: false
          schema:
            type: string
          example: '1'
      responses:
        '200':
          description: List of transportation orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransportationOrder'
              example:
                - id: 472
                  order_number: '#00472'
                  customer_id: 1
                  status: DELIVERED
                  amount: '43.19'
                  order_date: '2025-11-25T22:31:47.000Z'
                  expected_delivery: '2025-12-02T00:00:00.000Z'
                  shipping_address: 6806 East Landchester, Plano, Utah 10652
                  shipping_city: Plano
                  shipping_state: Utah
                  shipping_zip_code: '10652'
                  shipping_method: Standard Delivery
                  tracking_number: SR0472
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transportation-orders/{id}/driver:
    put:
      summary: Assign a driver to a transportation order
      description: Assigns an existing driver to the specified transportation order. Replaces any previously assigned driver. The operation is idempotent.
      operationId: assignDriverToOrder
      tags:
        - TransportationOrders
      parameters:
        - $ref: '#/components/parameters/PathIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignDriverInput'
            example:
              driver_id: 7
      responses:
        '200':
          description: Driver assigned successfully.
        '400':
          description: Request payload has an invalid structure or contains missing/invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: driver_id is required and must be a positive integer
        '404':
          description: Transportation order or driver with the given ID does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Transportation order not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notifications:
    get:
      summary: List notifications by user
      description: Returns a paginated list of notifications for the given user. Results are ordered by created_at descending.
      operationId: getNotificationsByUserId
      tags:
        - Notifications
      parameters:
        - name: userId
          in: query
          description: User ID to fetch notifications for. Required.
          required: true
          schema:
            type: string
            pattern: ^[1-9]\d*$
          example: '1'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
              example:
                data:
                  - id: 115
                    user_id: 1
                    type: warning
                    message: Low fuel alert for assigned truck
                    created_at: '2026-03-13T19:10:40.000Z'
                    is_read: false
                  - id: 14
                    user_id: 1
                    type: warning
                    message: Low fuel alert for assigned truck
                    created_at: '2026-03-12T22:49:04.000Z'
                    is_read: false
                  - id: 77
                    user_id: 1
                    type: success
                    message: Payment received successfully
                    created_at: '2026-03-11T21:38:45.000Z'
                    is_read: true
                pagination:
                  page: 1
                  limit: 3
                  total: 4
                  totalPages: 2
        '400':
          description: The userId query parameter is missing or invalid (must be a positive integer).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      required: false
      schema:
        type: string
        default: '1'
      example: '1'
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (max 100)
      required: false
      schema:
        type: string
        default: '20'
      example: '20'
    PathIdParam:
      name: id
      in: path
      description: Numeric resource identifier
      required: true
      schema:
        type: integer
      example: 1
  responses:
    BadRequest:
      description: The request contains unknown or invalid query parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: An unexpected server-side error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Pagination:
      type: object
      description: Pagination metadata attached to list responses.
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 20
        total:
          type: integer
          description: Total number of records across all pages
          example: 50
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
    ErrorResponse:
      type: object
      description: Standard error envelope returned on 4xx / 5xx responses.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: Vehicle not found
...
    DriverListItem:
      type: object
      description: A single driver record as returned in list and create responses.
      required:
        - id
      properties:
        id:
          type: integer
          description: Unique driver identifier
          example: 1
        first_name:
          type: string
          nullable: true
          maxLength: 50
          description: Driver first name
          example: Matt
        last_name:
          type: string
          nullable: true
          maxLength: 50
          description: Driver last name
          example: Jerde
        email:
          type: string
          nullable: true
          maxLength: 100
          description: Driver email address
          example: celiawalker@swift.net
        phone:
          type: string
          nullable: true
          maxLength: 20
          description: Driver phone number
          example: '4111234962'
        contract_type:
          type: string
          nullable: true
          maxLength: 20
          description: Contract type (e.g. CONTRACTOR, FULL_TIME)
          example: CONTRACTOR
        status:
          type: string
          nullable: true
          maxLength: 20
          description: Driver status (e.g. ON_ROUTE, ACTIVE, RESTING)
          example: ON_ROUTE
    DriverCreateInput:
      type: object
      description: Payload for creating a new driver. All fields are optional.
      properties:
        first_name:
          type: string
          nullable: true
          maxLength: 50
          description: Driver first name
          example: Jan
        last_name:
          type: string
          nullable: true
          maxLength: 50
          description: Driver last name
          example: Kowalski
        email:
          type: string
          nullable: true
          maxLength: 100
          description: Driver email address
          example: jan.kowalski@example.com
        phone:
          type: string
          nullable: true
          maxLength: 20
          description: Driver phone number
          example: '1234567890'
        contract_type:
          type: string
          nullable: true
          maxLength: 20
          description: Contract type (e.g. FULL_TIME, CONTRACTOR)
          example: FULL_TIME
        status:
          type: string
          nullable: true
          maxLength: 20
          description: Driver status
          example: ACTIVE
    DriverLicense:
      type: object
      description: Driver license record with license type details (returned in driver detail).
      properties:
        id:
          type: integer
          description: License record identifier
          example: 1
        document_number:
          type: string
          nullable: true
          description: License document number
          example: CXX232105
        issue_date:
          type: string
          format: date
          nullable: true
          description: License issue date (YYYY-MM-DD)
          example: '2020-08-30'
        expiry_date:
          type: string
          format: date
          nullable: true
          description: License expiry date (YYYY-MM-DD)
          example: '2027-09-21'
        status:
          type: string
          nullable: true
          description: License status (e.g. active)
          example: active
        code:
          type: string
          nullable: true
          description: License type code
          example: C
        name:
          type: string
          nullable: true
          description: License type name
          example: Prawo jazdy kat. C
        description:
          type: string
          nullable: true
          description: License type description
          example: Pojazdy ciężarowe powyżej 3,5t
    DriverDetail:
      type: object
      description: Full driver record including associated licenses (returned by GET /drivers/{id}).
      required:
        - id
        - licenses
      properties:
        id:
          type: integer
          description: Unique driver identifier
          example: 1
        first_name:
          type: string
          nullable: true
          maxLength: 50
          description: Driver first name
          example: Matt
        last_name:
          type: string
          nullable: true
          maxLength: 50
          description: Driver last name
          example: Jerde
        email:
          type: string
          nullable: true
          maxLength: 100
          description: Driver email address
          example: celiawalker@swift.net
        phone:
          type: string
          nullable: true
          maxLength: 20
          description: Driver phone number
          example: '4111234962'
        contract_type:
          type: string
          nullable: true
          maxLength: 20
          description: Contract type (e.g. CONTRACTOR, FULL_TIME)
          example: CONTRACTOR
        status:
          type: string
          nullable: true
          maxLength: 20
          description: Driver status
          example: ON_ROUTE
        licenses:
          type: array
          description: Associated driver licenses with type details
          items:
            $ref: '#/components/schemas/DriverLicense'
    TransportationOrder:
      type: object
      description: A single transportation order as returned by the API.
      required:
        - id
        - order_number
        - customer_id
        - status
        - amount
        - order_date
      properties:
        id:
          type: integer
          description: Unique order identifier
          example: 472
        order_number:
          type: string
          maxLength: 20
          description: Human-readable order number (e.g.
          example: '#00472'
        customer_id:
          type: integer
          description: ID of the customer who placed the order
          example: 1
        driver_id:
          type: integer
          nullable: true
          description: ID of the driver assigned to this order. Null when no driver has been assigned yet.
          example: 7
        status:
          type: string
          maxLength: 20
          description: Order status (e.g. DELIVERED, PENDING)
          example: DELIVERED
        amount:
          type: string
          description: |
            Order total amount. Returned as a decimal string (e.g. "43.19") because Postgres numeric is serialised as a string by the pg driver.
          example: '43.19'
        order_date:
          type: string
          format: date-time
          description: When the order was placed (ISO 8601)
          example: '2025-11-25T22:31:47.000Z'
        expected_delivery:
          type: string
          format: date-time
          nullable: true
          description: Expected delivery date (ISO 8601; Postgres date may be serialised with time component)
          example: '2025-12-02T00:00:00.000Z'
        shipping_address:
          type: string
          maxLength: 255
          nullable: true
          description: Delivery street address
          example: 6806 East Landchester, Plano, Utah 10652
        shipping_city:
          type: string
          maxLength: 100
          nullable: true
          description: Delivery city
          example: Plano
        shipping_state:
          type: string
          maxLength: 50
          nullable: true
          description: Delivery state or region
          example: Utah
        shipping_zip_code:
          type: string
          maxLength: 20
          nullable: true
          description: Delivery postal code
          example: '10652'
        shipping_method:
          type: string
          maxLength: 50
          nullable: true
          description: Shipping method name
          example: Standard Delivery
        tracking_number:
          type: string
          maxLength: 50
          nullable: true
          description: Carrier tracking number
          example: SR0472
    AssignDriverInput:
      type: object
      description: Payload for assigning a driver to a transportation order.
      required:
        - driver_id
      properties:
        driver_id:
          type: integer
          description: ID of the driver to assign to the order.
          example: 7
    Notification:
      type: object
      description: A single notification as returned by the API.
      required:
        - id
        - user_id
        - type
        - message
        - created_at
        - is_read
      properties:
        id:
          type: integer
          description: Unique notification identifier
          example: 115
        user_id:
          type: integer
          description: ID of the user who receives the notification
          example: 1
        type:
          type: string
          maxLength: 20
          description: Notification type (e.g. warning, success)
          example: warning
        message:
          type: string
          description: Notification message content
          example: Low fuel alert for assigned truck
        created_at:
          type: string
          format: date-time
          description: When the notification was created (ISO 8601)
          example: '2026-03-13T19:10:40.000Z'
        is_read:
          type: boolean
          description: Whether the notification has been read
          example: false
    NotificationListResponse:
      type: object
      description: Paginated list of notifications.
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        pagination:
          $ref: '#/components/schemas/Pagination'
...